Edit the update method in item class to include the manufacturer id, removed assigned ids for items and manufacturers in seed.rb

+++ db/shop.sql
@@ -6,7 +6,7 @@ CREATE TABLE manufacturers(
   id SERIAL4 PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
   address VARCHAR(255) NOT NULL,
-  telephone INT
+  telephone INT8
 );
 
 CREATE TABLE items(
